refactor: move diff file writers to output.diff
