+++ src/cja_auto_sdr/diff/writers.py
@@ -1,4 +1,4 @@
-"""Diff output writers (current implementation lives in generator for compatibility)."""
+"""Diff output writers (canonical implementation in output.diff; this module forwards for compatibility)."""
 
 __all__ = [
     "write_diff_console_output",
@@ -13,4 +13,4 @@ __all__ = [
 
 from cja_auto_sdr.core.lazy import make_getattr
 
-__getattr__ = make_getattr(__name__, __all__, target_module="cja_auto_sdr.generator")
+__getattr__ = make_getattr(__name__, __all__, target_module="cja_auto_sdr.output.diff")
