+++ src/cja_auto_sdr/generator.py
@@ -2639,8 +2639,13 @@ from cja_auto_sdr.output.diff import (
     _get_inventory_change_detail,
     detect_breaking_changes,
     write_diff_console_output,
+    write_diff_csv_output,
+    write_diff_excel_output,
     write_diff_grouped_by_field_output,
+    write_diff_html_output,
+    write_diff_json_output,
     write_diff_markdown_output,
+    write_diff_output,
     write_diff_pr_comment_output,
 )
 
@@ -2655,953 +2660,13 @@ from cja_auto_sdr.output.sdr import (
     write_markdown_output,
 )
 
-
-def write_diff_json_output(
-    diff_result: DiffResult,
-    base_filename: str,
-    output_dir: str | Path,
-    logger: logging.Logger,
-    changes_only: bool = False,
-) -> str:
-    """
-    Write diff comparison to JSON format.
-
-    Args:
-        diff_result: The DiffResult to output
-        base_filename: Base filename without extension
-        output_dir: Output directory path
-        logger: Logger instance
-        changes_only: Only include changed items
-
-    Returns:
-        Path to JSON output file
-    """
-    try:
-        logger.info("Generating diff JSON output...")
-
-        summary = diff_result.summary
-        meta = diff_result.metadata_diff
-
-        def serialize_component_diff(d: ComponentDiff) -> dict:
-            return {
-                "id": d.id,
-                "name": d.name,
-                "change_type": d.change_type.value,
-                "changed_fields": {k: {"source": v[0], "target": v[1]} for k, v in (d.changed_fields or {}).items()},
-                "source_data": d.source_data,
-                "target_data": d.target_data,
-            }
-
-        def serialize_inventory_diff(d: InventoryItemDiff) -> dict:
-            return {
-                "id": d.id,
-                "name": d.name,
-                "change_type": d.change_type.value,
-                "inventory_type": d.inventory_type,
-                "changed_fields": {k: {"source": v[0], "target": v[1]} for k, v in (d.changed_fields or {}).items()},
-                "source_data": d.source_data,
-                "target_data": d.target_data,
-            }
-
-        # Filter diffs if changes_only
-        metric_diffs = diff_result.metric_diffs
-        dimension_diffs = diff_result.dimension_diffs
-        if changes_only:
-            metric_diffs = [d for d in metric_diffs if d.change_type != ChangeType.UNCHANGED]
-            dimension_diffs = [d for d in dimension_diffs if d.change_type != ChangeType.UNCHANGED]
-
-        json_data = {
-            "metadata": {
-                "generated_at": diff_result.generated_at,
-                "tool_version": diff_result.tool_version,
-                "source_label": diff_result.source_label,
-                "target_label": diff_result.target_label,
-            },
-            "source": {
-                "id": meta.source_id,
-                "name": meta.source_name,
-                "owner": meta.source_owner,
-                "description": meta.source_description,
-            },
-            "target": {
-                "id": meta.target_id,
-                "name": meta.target_name,
-                "owner": meta.target_owner,
-                "description": meta.target_description,
-            },
-            "summary": {
-                "source_metrics_count": summary.source_metrics_count,
-                "target_metrics_count": summary.target_metrics_count,
-                "source_dimensions_count": summary.source_dimensions_count,
-                "target_dimensions_count": summary.target_dimensions_count,
-                "metrics_added": summary.metrics_added,
-                "metrics_removed": summary.metrics_removed,
-                "metrics_modified": summary.metrics_modified,
-                "metrics_unchanged": summary.metrics_unchanged,
-                "metrics_change_percent": summary.metrics_change_percent,
-                "dimensions_added": summary.dimensions_added,
-                "dimensions_removed": summary.dimensions_removed,
-                "dimensions_modified": summary.dimensions_modified,
-                "dimensions_unchanged": summary.dimensions_unchanged,
-                "dimensions_change_percent": summary.dimensions_change_percent,
-                "has_changes": summary.has_changes,
-                "total_changes": summary.total_changes,
-                "total_added": summary.total_added,
-                "total_removed": summary.total_removed,
-                "total_modified": summary.total_modified,
-                "total_summary": summary.total_summary,
-            },
-            "metric_diffs": [serialize_component_diff(d) for d in metric_diffs],
-            "dimension_diffs": [serialize_component_diff(d) for d in dimension_diffs],
-        }
-
-        # Add inventory diffs if present
-        if diff_result.has_inventory_diffs:
-            inventory_summary = {}
-            if summary.source_calc_metrics_count > 0 or summary.target_calc_metrics_count > 0:
-                inventory_summary["calculated_metrics"] = {
-                    "source_count": summary.source_calc_metrics_count,
-                    "target_count": summary.target_calc_metrics_count,
-                    "added": summary.calc_metrics_added,
-                    "removed": summary.calc_metrics_removed,
-                    "modified": summary.calc_metrics_modified,
-                    "unchanged": summary.calc_metrics_unchanged,
-                }
-            if summary.source_segments_count > 0 or summary.target_segments_count > 0:
-                inventory_summary["segments"] = {
-                    "source_count": summary.source_segments_count,
-                    "target_count": summary.target_segments_count,
-                    "added": summary.segments_added,
-                    "removed": summary.segments_removed,
-                    "modified": summary.segments_modified,
-                    "unchanged": summary.segments_unchanged,
-                }
-            json_data["inventory_summary"] = inventory_summary
-
-            if diff_result.calc_metrics_diffs is not None:
-                calc_diffs = diff_result.calc_metrics_diffs
-                if changes_only:
-                    calc_diffs = [d for d in calc_diffs if d.change_type != ChangeType.UNCHANGED]
-                json_data["calculated_metrics_diffs"] = [serialize_inventory_diff(d) for d in calc_diffs]
-
-            if diff_result.segments_diffs is not None:
-                seg_diffs = diff_result.segments_diffs
-                if changes_only:
-                    seg_diffs = [d for d in seg_diffs if d.change_type != ChangeType.UNCHANGED]
-                json_data["segments_diffs"] = [serialize_inventory_diff(d) for d in seg_diffs]
-
-        json_file = os.path.join(output_dir, f"{base_filename}.json")
-        with open(json_file, "w", encoding="utf-8") as f:
-            json.dump(json_data, f, indent=2, ensure_ascii=False)
-
-        logger.info(f"Diff JSON file created: {json_file}")
-        return json_file
-
-    except (OSError, TypeError, ValueError) as e:
-        logger.error(_format_error_msg("creating diff JSON file", error=e))
-        raise
-
-
-def write_diff_html_output(
-    diff_result: DiffResult,
-    base_filename: str,
-    output_dir: str | Path,
-    logger: logging.Logger,
-    changes_only: bool = False,
-) -> str:
-    """
-    Write diff comparison to HTML format with professional styling.
-
-    Args:
-        diff_result: The DiffResult to output
-        base_filename: Base filename without extension
-        output_dir: Output directory path
-        logger: Logger instance
-        changes_only: Only include changed items
-
-    Returns:
-        Path to HTML output file
-    """
-    try:
-        logger.info("Generating diff HTML output...")
-
-        _html_escape = html.escape  # Capture before nested functions shadow 'html'
-
-        def _safe_html_escape(value: Any) -> str:
-            """Escape arbitrary values safely for HTML rendering."""
-            return _html_escape("" if value is None else str(value))
-
-        summary = diff_result.summary
-        meta = diff_result.metadata_diff
-        html_parts = []
-
-        # HTML header with CSS
-        html_parts.append("""
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Data View Comparison Report</title>
-    <style>
-        body {
-            font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-            margin: 0;
-            padding: 20px;
-            background-color: #f5f5f5;
-            color: #333;
-        }
-        .container {
-            max-width: 1400px;
-            margin: 0 auto;
-            background-color: white;
-            padding: 30px;
-            box-shadow: 0 0 10px rgba(0,0,0,0.1);
-            border-radius: 8px;
-        }
-        h1 {
-            color: #2c3e50;
-            border-bottom: 3px solid #3498db;
-            padding-bottom: 10px;
-        }
-        h2 {
-            color: #34495e;
-            margin-top: 30px;
-            border-left: 4px solid #3498db;
-            padding-left: 15px;
-        }
-        .metadata {
-            background-color: #ecf0f1;
-            padding: 15px;
-            border-radius: 5px;
-            margin-bottom: 20px;
-        }
-        .summary-table {
-            width: 100%;
-            border-collapse: collapse;
-            margin: 20px 0;
-        }
-        .summary-table th, .summary-table td {
-            padding: 12px;
-            text-align: left;
-            border-bottom: 1px solid #ddd;
-        }
-        .summary-table th {
-            background-color: #3498db;
-            color: white;
-        }
-        .diff-table {
-            width: 100%;
-            border-collapse: collapse;
-            margin: 20px 0;
-            font-size: 14px;
-        }
-        .diff-table th {
-            background-color: #34495e;
-            color: white;
-            padding: 12px;
-            text-align: left;
-        }
-        .diff-table td {
-            padding: 10px 12px;
-            border-bottom: 1px solid #ddd;
-        }
-        .row-added {
-            background-color: #d4edda !important;
-        }
-        .row-removed {
-            background-color: #f8d7da !important;
-        }
-        .row-modified {
-            background-color: #fff3cd !important;
-        }
-        .badge {
-            padding: 4px 8px;
-            border-radius: 4px;
-            font-weight: bold;
-            font-size: 12px;
-        }
-        .badge-added { background-color: #28a745; color: white; }
-        .badge-removed { background-color: #dc3545; color: white; }
-        .badge-modified { background-color: #ffc107; color: black; }
-        .no-changes {
-            color: #28a745;
-            font-weight: bold;
-            font-size: 18px;
-            text-align: center;
-            padding: 20px;
-        }
-        .total-changes {
-            font-size: 18px;
-            font-weight: bold;
-            margin: 20px 0;
-        }
-        .footer {
-            margin-top: 30px;
-            padding-top: 20px;
-            border-top: 1px solid #ddd;
-            text-align: center;
-            color: #7f8c8d;
-            font-size: 12px;
-        }
-        code {
-            background-color: #f8f9fa;
-            padding: 2px 6px;
-            border-radius: 3px;
-            font-family: 'Consolas', monospace;
-            font-size: 13px;
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <h1>Data View Comparison Report</h1>
-""")
-
-        # Metadata section
-        html_parts.append(f"""
-        <div class="metadata">
-            <p><strong>{_safe_html_escape(diff_result.source_label)}:</strong> {_safe_html_escape(meta.source_name)} (<code>{_safe_html_escape(meta.source_id)}</code>)</p>
-            <p><strong>{_safe_html_escape(diff_result.target_label)}:</strong> {_safe_html_escape(meta.target_name)} (<code>{_safe_html_escape(meta.target_id)}</code>)</p>
-            <p><strong>Generated:</strong> {_safe_html_escape(diff_result.generated_at)}</p>
-        </div>
-""")
-
-        # Summary table
-        html_parts.append(f"""
-        <h2>Summary</h2>
-        <table class="summary-table">
-            <tr>
-                <th>Component</th>
-                <th>{_safe_html_escape(diff_result.source_label)}</th>
-                <th>{_safe_html_escape(diff_result.target_label)}</th>
-                <th>Added</th>
-                <th>Removed</th>
-                <th>Modified</th>
-                <th>Unchanged</th>
-                <th>Changed</th>
-            </tr>
-            <tr>
-                <td>Metrics</td>
-                <td>{summary.source_metrics_count}</td>
-                <td>{summary.target_metrics_count}</td>
-                <td><span class="badge badge-added">+{summary.metrics_added}</span></td>
-                <td><span class="badge badge-removed">-{summary.metrics_removed}</span></td>
-                <td><span class="badge badge-modified">~{summary.metrics_modified}</span></td>
-                <td>{summary.metrics_unchanged}</td>
-                <td>{summary.metrics_change_percent:.1f}%</td>
-            </tr>
-            <tr>
-                <td>Dimensions</td>
-                <td>{summary.source_dimensions_count}</td>
-                <td>{summary.target_dimensions_count}</td>
-                <td><span class="badge badge-added">+{summary.dimensions_added}</span></td>
-                <td><span class="badge badge-removed">-{summary.dimensions_removed}</span></td>
-                <td><span class="badge badge-modified">~{summary.dimensions_modified}</span></td>
-                <td>{summary.dimensions_unchanged}</td>
-                <td>{summary.dimensions_change_percent:.1f}%</td>
-            </tr>
-        </table>
-""")
-
-        # Add inventory rows to summary if present
-        if diff_result.has_inventory_diffs:
-            inv_rows = []
-            if summary.source_calc_metrics_count > 0 or summary.target_calc_metrics_count > 0:
-                inv_rows.append(f"""
-            <tr>
-                <td><strong>Calc Metrics</strong></td>
-                <td>{summary.source_calc_metrics_count}</td>
-                <td>{summary.target_calc_metrics_count}</td>
-                <td><span class="badge badge-added">+{summary.calc_metrics_added}</span></td>
-                <td><span class="badge badge-removed">-{summary.calc_metrics_removed}</span></td>
-                <td><span class="badge badge-modified">~{summary.calc_metrics_modified}</span></td>
-                <td>{summary.calc_metrics_unchanged}</td>
-                <td>-</td>
-            </tr>""")
-            if summary.source_segments_count > 0 or summary.target_segments_count > 0:
-                inv_rows.append(f"""
-            <tr>
-                <td><strong>Segments</strong></td>
-                <td>{summary.source_segments_count}</td>
-                <td>{summary.target_segments_count}</td>
-                <td><span class="badge badge-added">+{summary.segments_added}</span></td>
-                <td><span class="badge badge-removed">-{summary.segments_removed}</span></td>
-                <td><span class="badge badge-modified">~{summary.segments_modified}</span></td>
-                <td>{summary.segments_unchanged}</td>
-                <td>-</td>
-            </tr>""")
-            if inv_rows:
-                # Insert inventory rows before the closing </table> tag
-                html_parts[-1] = html_parts[-1].replace("</table>", "".join(inv_rows) + "\n        </table>")
-
-        if not summary.has_changes:
-            html_parts.append('<p class="no-changes">No differences found.</p>')
-        else:
-            html_parts.append(f'<p class="total-changes">Total changes: {summary.total_changes}</p>')
-
-        # Helper function to generate diff table
-        def generate_diff_table(diffs: list[ComponentDiff], title: str):
-            changes = [d for d in diffs if d.change_type != ChangeType.UNCHANGED]
-            if not changes and changes_only:
-                return ""
-
-            html = f"<h2>{title}</h2>\n"
-            if not changes:
-                html += "<p><em>No changes</em></p>\n"
-                return html
-
-            html += """<table class="diff-table">
-                <tr>
-                    <th>Status</th>
-                    <th>ID</th>
-                    <th>Name</th>
-                    <th>Details</th>
-                </tr>"""
-
-            for diff in changes:
-                row_class = f"row-{diff.change_type.value}"
-                badge_class = f"badge-{diff.change_type.value}"
-                badge_text = diff.change_type.value.upper()
-                detail = _get_change_detail(diff)
-                detail_escaped = _html_escape(detail)
-
-                html += f"""
-                <tr class="{row_class}">
-                    <td><span class="badge {badge_class}">{badge_text}</span></td>
-                    <td><code>{_html_escape(str(diff.id))}</code></td>
-                    <td>{_html_escape(str(diff.name))}</td>
-                    <td>{detail_escaped}</td>
-                </tr>"""
-
-            html += "</table>\n"
-            return html
-
-        html_parts.append(generate_diff_table(diff_result.metric_diffs, "Metrics Changes"))
-        html_parts.append(generate_diff_table(diff_result.dimension_diffs, "Dimensions Changes"))
-
-        # Inventory diff sections (if present)
-        if diff_result.has_inventory_diffs:
-
-            def generate_inventory_diff_table(diffs: list[InventoryItemDiff] | None, title: str):
-                if diffs is None:
-                    return ""
-                changes = [d for d in diffs if d.change_type != ChangeType.UNCHANGED]
-                if not changes and changes_only:
-                    return ""
-
-                html = f"<h2>{title}</h2>\n"
-                if not changes:
-                    html += "<p><em>No changes</em></p>\n"
-                    return html
-
-                html += """<table class="diff-table">
-                    <tr>
-                        <th>Status</th>
-                        <th>ID</th>
-                        <th>Name</th>
-                        <th>Details</th>
-                    </tr>"""
-
-                for diff in changes:
-                    row_class = f"row-{diff.change_type.value}"
-                    badge_class = f"badge-{diff.change_type.value}"
-                    badge_text = diff.change_type.value.upper()
-                    detail = _get_inventory_change_detail(diff)
-                    detail_escaped = _html_escape(detail)
-
-                    html += f"""
-                    <tr class="{row_class}">
-                        <td><span class="badge {badge_class}">{badge_text}</span></td>
-                        <td><code>{_html_escape(str(diff.id))}</code></td>
-                        <td>{_html_escape(str(diff.name))}</td>
-                        <td>{detail_escaped}</td>
-                    </tr>"""
-
-                html += "</table>\n"
-                return html
-
-            html_parts.append(
-                "<h2 style='border-top: 2px solid #3498db; padding-top: 20px; margin-top: 30px;'>Inventory Changes</h2>",
-            )
-            html_parts.append(
-                generate_inventory_diff_table(diff_result.calc_metrics_diffs, "Calculated Metrics Changes"),
-            )
-            html_parts.append(generate_inventory_diff_table(diff_result.segments_diffs, "Segments Changes"))
-
-        # Footer
-        html_parts.append(f"""
-        <div class="footer">
-            <p>Generated by CJA SDR Generator v{diff_result.tool_version}</p>
-        </div>
-    </div>
-</body>
-</html>
-""")
-
-        html_file = os.path.join(output_dir, f"{base_filename}.html")
-        with open(html_file, "w", encoding="utf-8") as f:
-            f.write("\n".join(html_parts))
-
-        logger.info(f"Diff HTML file created: {html_file}")
-        return html_file
-
-    except (OSError, KeyError, TypeError, ValueError) as e:
-        logger.error(_format_error_msg("creating diff HTML file", error=e))
-        raise
-
-
-def write_diff_excel_output(
-    diff_result: DiffResult,
-    base_filename: str,
-    output_dir: str | Path,
-    logger: logging.Logger,
-    changes_only: bool = False,
-) -> str:
-    """
-    Write diff comparison to Excel format with color-coded rows.
-
-    Args:
-        diff_result: The DiffResult to output
-        base_filename: Base filename without extension
-        output_dir: Output directory path
-        logger: Logger instance
-        changes_only: Only include changed items
-
-    Returns:
-        Path to Excel output file
-    """
-    try:
-        logger.info("Generating diff Excel output...")
-
-        summary = diff_result.summary
-        meta = diff_result.metadata_diff
-        excel_file = os.path.join(output_dir, f"{base_filename}.xlsx")
-
-        with pd.ExcelWriter(excel_file, engine="xlsxwriter") as writer:
-            workbook = writer.book
-
-            # Define formats
-            added_format = workbook.add_format({"bg_color": "#d4edda", "border": 1})
-            removed_format = workbook.add_format({"bg_color": "#f8d7da", "border": 1})
-            modified_format = workbook.add_format({"bg_color": "#fff3cd", "border": 1})
-            normal_format = workbook.add_format({"border": 1})
-
-            # Summary sheet - build rows dynamically
-            summary_rows = [
-                {
-                    "Component": "Metrics",
-                    diff_result.source_label: summary.source_metrics_count,
-                    diff_result.target_label: summary.target_metrics_count,
-                    "Added": summary.metrics_added,
-                    "Removed": summary.metrics_removed,
-                    "Modified": summary.metrics_modified,
-                    "Unchanged": summary.metrics_unchanged,
-                    "Changed %": f"{summary.metrics_change_percent:.1f}%",
-                },
-                {
-                    "Component": "Dimensions",
-                    diff_result.source_label: summary.source_dimensions_count,
-                    diff_result.target_label: summary.target_dimensions_count,
-                    "Added": summary.dimensions_added,
-                    "Removed": summary.dimensions_removed,
-                    "Modified": summary.dimensions_modified,
-                    "Unchanged": summary.dimensions_unchanged,
-                    "Changed %": f"{summary.dimensions_change_percent:.1f}%",
-                },
-            ]
-
-            # Add inventory rows if present (check for actual inventory diffs)
-            if diff_result.calc_metrics_diffs is not None:
-                summary_rows.append(
-                    {
-                        "Component": "Calc Metrics",
-                        diff_result.source_label: summary.source_calc_metrics_count,
-                        diff_result.target_label: summary.target_calc_metrics_count,
-                        "Added": summary.calc_metrics_added,
-                        "Removed": summary.calc_metrics_removed,
-                        "Modified": summary.calc_metrics_modified,
-                        "Unchanged": summary.calc_metrics_unchanged,
-                        "Changed %": f"{summary.calc_metrics_change_percent:.1f}%",
-                    },
-                )
-            if diff_result.segments_diffs is not None:
-                summary_rows.append(
-                    {
-                        "Component": "Segments",
-                        diff_result.source_label: summary.source_segments_count,
-                        diff_result.target_label: summary.target_segments_count,
-                        "Added": summary.segments_added,
-                        "Removed": summary.segments_removed,
-                        "Modified": summary.segments_modified,
-                        "Unchanged": summary.segments_unchanged,
-                        "Changed %": f"{summary.segments_change_percent:.1f}%",
-                    },
-                )
-
-            summary_df = pd.DataFrame(summary_rows)
-            summary_df.to_excel(writer, sheet_name="Summary", index=False)
-
-            # Metadata sheet
-            metadata_data = {
-                "Property": [
-                    "Source ID",
-                    "Source Name",
-                    "Target ID",
-                    "Target Name",
-                    "Generated At",
-                    "Has Changes",
-                    "Total Changes",
-                ],
-                "Value": [
-                    meta.source_id,
-                    meta.source_name,
-                    meta.target_id,
-                    meta.target_name,
-                    diff_result.generated_at,
-                    str(summary.has_changes),
-                    summary.total_changes,
-                ],
-            }
-            metadata_df = pd.DataFrame(metadata_data)
-            metadata_df.to_excel(writer, sheet_name="Metadata", index=False)
-
-            # Helper function to write diff sheet
-            def write_diff_sheet(diffs: list[ComponentDiff], sheet_name: str):
-                if changes_only:
-                    diffs = [d for d in diffs if d.change_type != ChangeType.UNCHANGED]
-
-                if not diffs:
-                    df = pd.DataFrame({"Message": ["No changes"]})
-                    df.to_excel(writer, sheet_name=sheet_name, index=False)
-                    return
-
-                rows = [
-                    {
-                        "Status": diff.change_type.value.upper(),
-                        "ID": diff.id,
-                        "Name": diff.name,
-                        "Details": _get_change_detail(diff),
-                    }
-                    for diff in diffs
-                ]
-
-                df = pd.DataFrame(rows)
-                df.to_excel(writer, sheet_name=sheet_name, index=False)
-
-                # Apply color formatting
-                worksheet = writer.sheets[sheet_name]
-                for row_idx, diff in enumerate(diffs, start=1):
-                    if diff.change_type == ChangeType.ADDED:
-                        fmt = added_format
-                    elif diff.change_type == ChangeType.REMOVED:
-                        fmt = removed_format
-                    elif diff.change_type == ChangeType.MODIFIED:
-                        fmt = modified_format
-                    else:
-                        fmt = normal_format
-
-                    for col_idx in range(len(df.columns)):
-                        worksheet.write(row_idx, col_idx, df.iloc[row_idx - 1, col_idx], fmt)
-
-            write_diff_sheet(diff_result.metric_diffs, "Metrics Diff")
-            write_diff_sheet(diff_result.dimension_diffs, "Dimensions Diff")
-
-            # Helper function to write inventory diff sheet
-            def write_inventory_diff_sheet(diffs: list[InventoryItemDiff] | None, sheet_name: str):
-                if diffs is None:
-                    return
-
-                if changes_only:
-                    diffs = [d for d in diffs if d.change_type != ChangeType.UNCHANGED]
-
-                if not diffs:
-                    df = pd.DataFrame({"Message": ["No changes"]})
-                    df.to_excel(writer, sheet_name=sheet_name, index=False)
-                    return
-
-                rows = [
-                    {
-                        "Status": diff.change_type.value.upper(),
-                        "ID": diff.id,
-                        "Name": diff.name,
-                        "Details": _get_inventory_change_detail(diff),
-                    }
-                    for diff in diffs
-                ]
-
-                df = pd.DataFrame(rows)
-                df.to_excel(writer, sheet_name=sheet_name, index=False)
-
-                # Apply color formatting
-                worksheet = writer.sheets[sheet_name]
-                for row_idx, diff in enumerate(diffs, start=1):
-                    if diff.change_type == ChangeType.ADDED:
-                        fmt = added_format
-                    elif diff.change_type == ChangeType.REMOVED:
-                        fmt = removed_format
-                    elif diff.change_type == ChangeType.MODIFIED:
-                        fmt = modified_format
-                    else:
-                        fmt = normal_format
-
-                    for col_idx in range(len(df.columns)):
-                        worksheet.write(row_idx, col_idx, df.iloc[row_idx - 1, col_idx], fmt)
-
-            # Write inventory diff sheets if present
-            if diff_result.calc_metrics_diffs is not None:
-                write_inventory_diff_sheet(diff_result.calc_metrics_diffs, "Calc Metrics Diff")
-            if diff_result.segments_diffs is not None:
-                write_inventory_diff_sheet(diff_result.segments_diffs, "Segments Diff")
-
-        logger.info(f"Diff Excel file created: {excel_file}")
-        return excel_file
-
-    except (OSError, KeyError, TypeError, ValueError) as e:
-        logger.error(_format_error_msg("creating diff Excel file", error=e))
-        raise
-
-
-def write_diff_csv_output(
-    diff_result: DiffResult,
-    base_filename: str,
-    output_dir: str | Path,
-    logger: logging.Logger,
-    changes_only: bool = False,
-) -> str:
-    """
-    Write diff comparison to CSV files.
-
-    Args:
-        diff_result: The DiffResult to output
-        base_filename: Base filename without extension
-        output_dir: Output directory path
-        logger: Logger instance
-        changes_only: Only include changed items
-
-    Returns:
-        Path to output directory containing CSV files
-    """
-    try:
-        logger.info("Generating diff CSV output...")
-
-        summary = diff_result.summary
-        meta = diff_result.metadata_diff
-
-        # Create subdirectory for CSV files
-        csv_dir = os.path.join(output_dir, f"{base_filename}_csv")
-        os.makedirs(csv_dir, exist_ok=True)
-
-        # Summary CSV - build rows dynamically
-        summary_rows = [
-            {
-                "Component": "Metrics",
-                "Source_Count": summary.source_metrics_count,
-                "Target_Count": summary.target_metrics_count,
-                "Added": summary.metrics_added,
-                "Removed": summary.metrics_removed,
-                "Modified": summary.metrics_modified,
-                "Unchanged": summary.metrics_unchanged,
-                "Changed_Percent": summary.metrics_change_percent,
-            },
-            {
-                "Component": "Dimensions",
-                "Source_Count": summary.source_dimensions_count,
-                "Target_Count": summary.target_dimensions_count,
-                "Added": summary.dimensions_added,
-                "Removed": summary.dimensions_removed,
-                "Modified": summary.dimensions_modified,
-                "Unchanged": summary.dimensions_unchanged,
-                "Changed_Percent": summary.dimensions_change_percent,
-            },
-        ]
-
-        # Add inventory rows if present (check for actual inventory diffs)
-        if diff_result.calc_metrics_diffs is not None:
-            summary_rows.append(
-                {
-                    "Component": "Calc_Metrics",
-                    "Source_Count": summary.source_calc_metrics_count,
-                    "Target_Count": summary.target_calc_metrics_count,
-                    "Added": summary.calc_metrics_added,
-                    "Removed": summary.calc_metrics_removed,
-                    "Modified": summary.calc_metrics_modified,
-                    "Unchanged": summary.calc_metrics_unchanged,
-                    "Changed_Percent": summary.calc_metrics_change_percent,
-                },
-            )
-        if diff_result.segments_diffs is not None:
-            summary_rows.append(
-                {
-                    "Component": "Segments",
-                    "Source_Count": summary.source_segments_count,
-                    "Target_Count": summary.target_segments_count,
-                    "Added": summary.segments_added,
-                    "Removed": summary.segments_removed,
-                    "Modified": summary.segments_modified,
-                    "Unchanged": summary.segments_unchanged,
-                    "Changed_Percent": summary.segments_change_percent,
-                },
-            )
-
-        pd.DataFrame(summary_rows).to_csv(os.path.join(csv_dir, "summary.csv"), index=False, encoding="utf-8")
-        logger.info("  Created: summary.csv")
-
-        # Metadata CSV
-        metadata_data = {
-            "Property": [
-                "source_id",
-                "source_name",
-                "target_id",
-                "target_name",
-                "generated_at",
-                "has_changes",
-                "total_changes",
-            ],
-            "Value": [
-                meta.source_id,
-                meta.source_name,
-                meta.target_id,
-                meta.target_name,
-                diff_result.generated_at,
-                str(summary.has_changes),
-                summary.total_changes,
-            ],
-        }
-        pd.DataFrame(metadata_data).to_csv(os.path.join(csv_dir, "metadata.csv"), index=False, encoding="utf-8")
-        logger.info("  Created: metadata.csv")
-
-        # Helper function to write diff CSV
-        def write_diff_csv(diffs: list[ComponentDiff], filename: str):
-            if changes_only:
-                diffs = [d for d in diffs if d.change_type != ChangeType.UNCHANGED]
-
-            rows = [
-                {
-                    "status": diff.change_type.value,
-                    "id": diff.id,
-                    "name": diff.name,
-                    "details": _get_change_detail(diff),
-                }
-                for diff in diffs
-            ]
-
-            pd.DataFrame(rows).to_csv(os.path.join(csv_dir, filename), index=False, encoding="utf-8")
-            logger.info(f"  Created: {filename}")
-
-        write_diff_csv(diff_result.metric_diffs, "metrics_diff.csv")
-        write_diff_csv(diff_result.dimension_diffs, "dimensions_diff.csv")
-
-        # Helper function to write inventory diff CSV
-        def write_inventory_diff_csv(diffs: list[InventoryItemDiff] | None, filename: str):
-            if diffs is None:
-                return
-
-            if changes_only:
-                diffs = [d for d in diffs if d.change_type != ChangeType.UNCHANGED]
-
-            rows = [
-                {
-                    "status": diff.change_type.value,
-                    "id": diff.id,
-                    "name": diff.name,
-                    "details": _get_inventory_change_detail(diff),
-                }
-                for diff in diffs
-            ]
-
-            pd.DataFrame(rows).to_csv(os.path.join(csv_dir, filename), index=False, encoding="utf-8")
-            logger.info(f"  Created: {filename}")
-
-        # Write inventory diff CSVs if present
-        if diff_result.calc_metrics_diffs is not None:
-            write_inventory_diff_csv(diff_result.calc_metrics_diffs, "calc_metrics_diff.csv")
-        if diff_result.segments_diffs is not None:
-            write_inventory_diff_csv(diff_result.segments_diffs, "segments_diff.csv")
-
-        logger.info(f"Diff CSV files created in: {csv_dir}")
-        return csv_dir
-
-    except (OSError, KeyError, TypeError, ValueError) as e:
-        logger.error(_format_error_msg("creating diff CSV files", error=e))
-        raise
-
-
-def write_diff_output(
-    diff_result: DiffResult,
-    output_format: str,
-    base_filename: str,
-    output_dir: str | Path,
-    logger: logging.Logger,
-    changes_only: bool = False,
-    summary_only: bool = False,
-    side_by_side: bool = False,
-    use_color: bool = True,
-    group_by_field: bool = False,
-    group_by_field_limit: int = 10,
-) -> str | None:
-    """
-    Write diff comparison output in specified format(s).
-
-    Args:
-        diff_result: The DiffResult to output
-        output_format: Output format ('console', 'json', 'markdown', 'html', 'excel', 'csv', 'all', 'pr-comment')
-        base_filename: Base filename without extension
-        output_dir: Output directory path
-        logger: Logger instance
-        changes_only: Only include changed items
-        summary_only: Only show summary (console only)
-        side_by_side: Show side-by-side comparison for modified items
-        use_color: Use ANSI color codes in console output
-        group_by_field: Group changes by field name instead of component
-        group_by_field_limit: Max items per section in group-by-field output (0 = unlimited)
-
-    Returns:
-        Console output string (for console/pr-comment format) or None
-    """
-    os.makedirs(output_dir, exist_ok=True)
-    output_files = []
-    console_output = None
-
-    # Handle group-by-field output mode
-    if group_by_field and should_generate_format(output_format, "console"):
-        console_output = write_diff_grouped_by_field_output(diff_result, use_color, group_by_field_limit)
-        print(console_output)
-        if output_format == "console":
-            return console_output
-
-    # Handle PR comment format
-    if output_format == "pr-comment":
-        console_output = write_diff_pr_comment_output(diff_result, changes_only)
-        print(console_output)
-        return console_output
-
-    if should_generate_format(output_format, "console") and not group_by_field:
-        console_output = write_diff_console_output(diff_result, changes_only, summary_only, side_by_side, use_color)
-        print(console_output)
-
-    if output_format == "console":
-        return console_output
-
-    if should_generate_format(output_format, "json"):
-        output_files.append(write_diff_json_output(diff_result, base_filename, output_dir, logger, changes_only))
-
-    if should_generate_format(output_format, "markdown"):
-        output_files.append(
-            write_diff_markdown_output(diff_result, base_filename, output_dir, logger, changes_only, side_by_side),
-        )
-
-    if should_generate_format(output_format, "html"):
-        output_files.append(write_diff_html_output(diff_result, base_filename, output_dir, logger, changes_only))
-
-    if should_generate_format(output_format, "excel"):
-        output_files.append(write_diff_excel_output(diff_result, base_filename, output_dir, logger, changes_only))
-
-    if should_generate_format(output_format, "csv"):
-        output_files.append(write_diff_csv_output(diff_result, base_filename, output_dir, logger, changes_only))
-
-    return console_output
-
+# ==================== DIFF FILE WRITERS (moved to output/diff/) ====================
+# - write_diff_json_output -> output.diff.json
+# - write_diff_html_output -> output.diff.html
+# - write_diff_excel_output -> output.diff.excel
+# - write_diff_csv_output -> output.diff.csv
+# - write_diff_output -> output.diff.__init__
+# (imported above from cja_auto_sdr.output.diff)
 
 # ==================== INVENTORY SUMMARY MODE ====================
